iio: ad9467: Mark sample rate property for JESD204 converters as read-only
The ad9467 driver allows to re-configure the input clock rate of a device
by changing the sampling_frequency attribute.

This works fine on LVDS based links where the converter is the master clock
source of the FPGA receiver.

It does not work for JESD204 based links though, which require changing
additional clocks as well as re-initializing the link after the clocks have
been updated. Changing just the converter clock will result in a broken
link that requires a system reboot to recover.

So until full JESD204 link re-initialization support is available make the
sampling_frequency property read-only.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>