iio: ad9467: Enable JESD204 link after AD9680 and lane rate are configured
For proper glitch free link startup the order in which things are enabled
and configured need to be considered.

Before the receiver side of the link is enabled the lane rate should be
configured since changing the lane rate while the link is active will cause
glitches.

In addition before the receiver side of the link is enabled the transmitter
side of the AD9680 should be up and running. This is necessary so that the
CDR on the receiver side has something to lock onto. Otherwise the CDRs
will be freerunning and each lane will run at a different clock rate, which
will cause a drift and very large latency variation.

Also make sure that the link is only enabled once (both conv->clk and
jesd_clk point to the same link clock).

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>