iio: ad9144: Add support for all JESD204 modes
The ad9144 driver currently hardcodes the register writes to setup the
device to the so called Mode 4 (M=2, L=4). No other modes are supported.

Add the infrastructure that is required to configure the device in other
modes.

For the time being still a static configuration is used. But this will make
it a lot easier to add full dynamic reconfiguration eventually.

It also makes it a lot easier to modify the driver to operate in different
modes.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>