iio: frequency: adf4360: Introduce adf4360_sync_config
Introduce adf4360_sync_config for synchronizing register configuration
with hardware. The sync_config function introduces a change in register
write behavior where a sync may be unconditional (all) or just registers
that have changed since the last sync. Previously, all three registers
were sent to hardware with an interval sleep before writing to the R
counter latch. Inspection of the data sheet clarifies that this delay is
only required at power up. The new behavior is to add the required power
up interval only when powering up or on probe otherwise only update
registers that have changed. If the charge pump current gets modified,
an interval will be required but there is currently no facility to
change the CPI other that the device tree which is accounted for on
probe. The result is faster tuning e.g a 15 MHz SPI interface requires
10ms to 20 ms for initial power up sync and 4.8 us frequency updates
kernel overhead not with standing.

Signed-off-by: Edward Kigwana <ekigwana@scires.com>